chore(DIST-242): Use Jarvis to deploy
Install Jarvis package (https://github.com/Typeform/jarvis) in Travis deploy job and use it for deployment to AWS. Since Jarvis is a private package we do not add it to dependencies in package.json or the installation for people from outside of the company would fail. It is installed only in the Travis CI deploy job and changes are not committed.